==> notebooks/1_baseline.ipynb <==
import pickle
import numpy as np
import pandas as pd

from collections import Counter

from sklearn.metrics import f1_score, make_scorer, confusion_matrix, \
    classification_report
from sklearn.base import BaseEstimator, ClassifierMixin
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB, BernoulliNB
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, cross_val_score, \
    StratifiedShuffleSplit, RandomizedSearchCV

# Load learn data

with open("../Data/Learn/labels.pkl", "rb") as f:
    learn_labels = pickle.load(f)

with open("../Data/Learn/sentences.pkl", "rb") as f:
    learn_sentences = pickle.load(f)

with open("../Data/Learn/sequences.pkl", "rb") as f:
    learn_sequences = pickle.load(f)

with open("../Data/dict.pkl", "rb") as f:
    vocabulary = pickle.load(f)
    reverse_vocabulary = {v: k for k, v in vocabulary.items()}

# Data exploration

def decode_sequence(sequence, decoding_dict):
    return " ".join((decoding_dict[x] for x in sequence))

print(len(vocabulary), len(reverse_vocabulary))
print(min(reverse_vocabulary.keys()), max(reverse_vocabulary.keys()))
sorted(vocabulary.keys(), key=len, reverse=True)[:5]

Counter(learn_labels)

print(len(learn_sentences))
learn_sentences[:5]

print(len(learn_sequences))
[decode_sequence(x, reverse_vocabulary) for x in learn_sequences][:5]

# Data splitting

X_train, X_test, y_train, y_test = train_test_split(
    learn_sequences, learn_labels, test_size=0.3,
    shuffle=True, stratify=learn_labels, random_state=42
)
print(len(X_train), len(y_train), len(X_test), len(y_test))
Counter(y_train), Counter(y_test)

# Utils

def f1(y_true, y_pred):
    return f1_score(y_true, y_pred, average="macro")

scoring = make_scorer(f1)

def score_model(model):
    model.fit(X_train, y_train)
    return scoring(model, X_test, y_test)


def report_model(model):
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred, digits=3))


def grid_search(model, param_grid, n_iter):
    rs = RandomizedSearchCV(
        model, param_grid, 
        n_iter=n_iter, 
        scoring=scoring,
        cv=StratifiedShuffleSplit(10, test_size=0.3, random_state=42),
        n_jobs=-1,
        random_state=42,
    )
    rs.fit(X_train, y_train)
    return pd.DataFrame(rs.cv_results_).sort_values(
        "mean_test_score", ascending=False)


def tokens_to_str(tokens):
    return list(map(str, tokens))

# Baseline models

### Majority

class Model(BaseEstimator, ClassifierMixin):
    def fit(self, X, y):
        return self
    def predict(self, X):
        return np.full(len(X), 'C')

"Test score: %s" % score_model(Model())

### N-grams + Naive Bayes

Model = lambda: Pipeline([
    ("vectorizer", CountVectorizer(lowercase=False, tokenizer=tokens_to_str)),
    ("classifier", MultinomialNB()),
])

results = grid_search(Model(), {
    "vectorizer__ngram_range": [(1, 1), (1, 2), (2, 2), (1, 3), (2, 3), (3, 3)],
    "classifier__alpha": np.logspace(-3, 0, 10),
}, n_iter=60)

best_params = results.params.iloc[0]
print(best_params)
"Test score: %s" % score_model(Model().set_params(**best_params))

print(best_params)
report_model(Model().set_params(**best_params))
