==> notebooks/2_dnn.ipynb <==
import os
import pickle
import numpy as np
import pandas as pd

from collections import Counter

from sklearn.metrics import f1_score, make_scorer, confusion_matrix, \
    classification_report
from sklearn.base import BaseEstimator, ClassifierMixin
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB, BernoulliNB
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, cross_val_score, \
    StratifiedShuffleSplit, RandomizedSearchCV
from sklearn.preprocessing import LabelEncoder


import tensorflow as tf

tf.logging.set_verbosity(tf.logging.INFO)

%matplotlib inline

# Data

with open("../Data/Learn/labels.pkl", "rb") as f:
    learn_labels = pickle.load(f)

with open("../Data/Learn/sequences.pkl", "rb") as f:
    learn_sequences = pickle.load(f)

X_train, X_test, y_train, y_test = train_test_split(
    learn_sequences, learn_labels, test_size=0.3,
    shuffle=True, stratify=learn_labels, random_state=42
)

# Model

def tokens_to_str(tokens):
    return list(map(str, tokens))

def convert_sparse_matrix_to_sparse_tensor(X):
    coo = X.tocoo()
    indices = np.mat([coo.row, coo.col]).transpose()
    return tf.SparseTensorValue(indices, coo.data, coo.shape)

def train_input_fn(x, y, shuffle=True, batch_size=128, num_epochs=None):
    n = x.shape[0]
    if n < batch_size:
        raise Exception("batch_size bigger than input")
    epoch_size = n // batch_size
    
    def input_fn():
        x_data = convert_sparse_matrix_to_sparse_tensor(x)
        y_data = tf.convert_to_tensor(y)
        i = tf.train.range_input_producer(epoch_size, num_epochs, shuffle=False).dequeue()
        ret_x = tf.sparse.slice(x_data, [i * batch_size, 0], [batch_size, x.shape[1]])
        ret_y = y_data[i * batch_size:(i + 1) * batch_size]
        return {"x": tf.sparse.to_dense(ret_x, validate_indices=False)}, ret_y
        
    return input_fn

def test_input_fn(x, y, batch_size=128):
    n = x.shape[0]
    if n < batch_size:
        raise Exception("batch_size bigger than input")
    epoch_size = n // batch_size
    
    def input_fn():
        x_data = convert_sparse_matrix_to_sparse_tensor(x)
        y_data = tf.convert_to_tensor(y)
        i = tf.train.range_input_producer(epoch_size, num_epochs=1, shuffle=False).dequeue()
        ret_x = tf.sparse.slice(x_data, [i * batch_size, 0], [batch_size, x.shape[1]])
        ret_y = y_data[i * batch_size:(i + 1) * batch_size]
        return {"x": tf.sparse.to_dense(ret_x, validate_indices=False)}, ret_y
        
    return input_fn

def score_model(model, X_test, y_test, batch_size=128):
    y_test = y_test[:X_test.shape[0] - X_test.shape[0] % batch_size]
    y_pred = np.array([x["class_ids"][0] for x in model.predict(
        test_input_fn(X_test, y_test, batch_size))])
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred, digits=3))
    return f1_score(y_test, y_pred, average="macro")

vectorizer = CountVectorizer(lowercase=False, tokenizer=tokens_to_str, ngram_range=(1, 2))
X_train_vect = vectorizer.fit_transform(X_train)
X_test_vect = vectorizer.transform(X_test)

vector_feature_column = tf.feature_column.numeric_column(key="x", shape=X_train_vect.shape[1])
classifier = tf.estimator.DNNClassifier(
    feature_columns=[vector_feature_column],
    hidden_units=[10],
    n_classes=2,
)

classifier.train(train_input_fn(X_train_vect, y_train, batch_size=128, num_epochs=10))

score_model(classifier, X_test_vect, y_test)

# Clean model

class DNNModel(BaseEstimator, ClassifierMixin):
    def __init__(self,
                 model_name=None,
                 checkpoints_dir="../checkpoints/",
                 hidden_units=(10,),
                 batch_size=128,
                 weight_class_M=1.0,
                 features_key="x",
                 weight_key="weight"
                 ):
        self.set_model_directory(checkpoints_dir, model_name)
        self.hidden_units = hidden_units
        self.batch_size = batch_size
        self.features_key = features_key
        self.weight_key = weight_key
        self.weight_class_M = weight_class_M

    def fit(self, X, y, num_epochs=1, warm_start=True):
        warm_start = self.check_warm_start(warm_start)
        if not warm_start:
            X, y = self.fit_and_apply_transformers(X, y)
            self.classifier_ = self.create_dnn_classifier()
        else:
            X, y = self.apply_transformers(X, y)

        self.classifier_.train(self.train_input_fn(X, y, num_epochs))
        return self

    def fit_and_apply_transformers(self, X, y):
        # Fit and transform X
        self.vectorizer_ = CountVectorizer(lowercase=False,
                                           tokenizer=self.tokens_to_str,
                                           ngram_range=(1, 2))
        X = self.vectorizer_.fit_transform(X)
        self.n_features_ = len(self.vectorizer_.vocabulary_)

        # Fit and transform y
        self.label_encoder_ = LabelEncoder()
        y = self.label_encoder_.fit_transform(y)
        self.n_classes_ = len(self.label_encoder_.classes_)

        return X, y

    def apply_transformers(self, X, y):
        X = self.vectorizer_.transform(X)
        y = self.label_encoder_.transform(y)
        return X, y

    def predict(self, X):
        X = self.vectorizer_.transform(X)
        predictions = list(self.classifier_.predict(self.predict_input_fn(X)))
        classes = np.array([p["class_ids"][0] for p in predictions])
        labels = self.label_encoder_.inverse_transform(classes)
        return labels

    def score(self, X, y, print_report=False):
        predictions = self.predict(X)
        if print_report:
            print(confusion_matrix(y, predictions))
            print(classification_report(y, predictions, digits=3))
        return f1_score(y, predictions, average="macro")

    def create_dnn_classifier(self):
        # Columns of X
        self.feature_columns_ = [tf.feature_column.numeric_column(
            key=self.features_key, shape=self.n_features_
        )]

        # Column of weights which will be used to compute loss
        weight_column = tf.feature_column.numeric_column(self.weight_key)

        # Model
        return tf.estimator.DNNClassifier(
            feature_columns=self.feature_columns_,
            weight_column=weight_column,
            hidden_units=self.hidden_units,
            n_classes=self.n_classes_,
            model_dir=self.model_dir
        )

    def train_input_fn(self, X, y, num_epochs):
        n = X.shape[0]
        num_batches = n // self.batch_size

        def input_fn():
            # Convert to tensors
            tf_X = self.convert_sparse_matrix_to_sparse_tensor(X)
            tf_y = tf.convert_to_tensor(y)

            # Batch iterator
            i = tf.train.range_input_producer(
                limit=num_batches, num_epochs=num_epochs, shuffle=False
            ).dequeue()

            # Slice sparse tensor using batch number, then convert to dense
            tf_X = tf.sparse.to_dense(tf.sparse.slice(
                tf_X, start=[i * self.batch_size, 0],
                size=[self.batch_size, self.n_features_]
            ), validate_indices=False)

            # Slice labels tensor using batch number
            output_y = tf_y[i * self.batch_size:(i + 1) * self.batch_size]

            # Create weights tensor using given weight of class "M"
            class_M = self.label_encoder_.transform(["M"])
            weights = tf.cast(tf.equal(output_y, class_M), tf.float64)
            weights = tf.multiply(weights, (self.weight_class_M - 1)) + 1

            # Return tensors in correct format
            return {self.features_key: tf_X, self.weight_key: weights}, output_y

        return input_fn

    def predict_input_fn(self, X):
        n = X.shape[0]

        # Ensure num_batches has the right value so that we predict every point
        if n % self.batch_size == 0:
            num_batches = n // self.batch_size
        else:
            num_batches = n // self.batch_size + 1

        def input_fn():
            # Convert to tensor
            tf_X = self.convert_sparse_matrix_to_sparse_tensor(X)

            # Batch iterator
            i = tf.train.range_input_producer(
                limit=num_batches, num_epochs=1, shuffle=False
            ).dequeue()

            # Slice sparse tensor using batch number, then convert to dense
            tf_X = tf.sparse.to_dense(tf.sparse.slice(
                tf_X, start=[i * self.batch_size, 0],
                size=[self.batch_size, self.n_features_]
            ), validate_indices=False)

            # Return tensor in correct format
            return {self.features_key: tf_X}

        return input_fn

    def check_warm_start(self, warm_start):
        if warm_start:
            # Check if model was already fitted
            try:
                self.classifier_
            except:
                warm_start = False
        return warm_start

    def set_model_directory(self, checkpoints_dir, model_name):
        if model_name is not None:
            self.model_dir = checkpoints_dir + model_name
            # Check model_dir doesn't already exist
            if os.path.exists(self.model_dir):
                raise ValueError("model_dir already exists")
        else:
            self.model_dir = None

    @staticmethod
    def tokens_to_str(tokens):
        # Used in CountVectorizer because we already have tokens
        return list(map(str, tokens))

    @staticmethod
    def convert_sparse_matrix_to_sparse_tensor(X):
        coo = X.tocoo()
        indices = np.mat([coo.row, coo.col]).transpose()
        return tf.SparseTensorValue(indices, coo.data, coo.shape)

num_epochs = 5
weight_class_M = Counter(y_train)["C"] / Counter(y_train)["M"]
range_units = [(10,), (25,), (50,), (100,), (250,), (500,), (1000,)]

for hidden_units in range_units:
    print("\n" + "-" * 80 + "\n")
    model = DNNModel(weight_class_M=weight_class_M, hidden_units=hidden_units)
    print(model)
    model.fit(X_train, y_train, num_epochs=num_epochs, warm_start=True)
    print("TEST F1-SCORE:", model.score(X_test, y_test, print_report=True))

num_epochs = 5
weight_class_M = Counter(y_train)["C"] / Counter(y_train)["M"]
range_units = [(100, 10), (100, 25), (100, 50), (100, 75), (100, 100)]

for hidden_units in range_units:
    print("\n" + "-" * 80 + "\n")
    model = DNNModel(weight_class_M=weight_class_M, hidden_units=hidden_units)
    print(model)
    model.fit(X_train, y_train, num_epochs=num_epochs, warm_start=True)
    print("TEST F1-SCORE:", model.score(X_test, y_test, print_report=True))

num_epochs = 10
weight_class_M = Counter(y_train)["C"] / Counter(y_train)["M"]
model = DNNModel(weight_class_M=weight_class_M, hidden_units=(250,))

for epoch in range(num_epochs):
    print("\n" + "-" * 80 + "\n" + "EPOCH %d\n" % epoch)
    model.fit(X_train, y_train, num_epochs=1, warm_start=True)
    print("TEST F1-SCORE:", model.score(X_test, y_test, print_report=True))

# Custom estimator

class DNNModel(BaseEstimator, ClassifierMixin):
    def __init__(self,
                 model_name=None,
                 checkpoints_dir="../checkpoints/",
                 hidden_units=(10,),
                 batch_size=128,
                 weight_class_M=1.0,
                 features_key="x",
                 weight_key="weight"
                 ):
        self.set_model_directory(checkpoints_dir, model_name)
        self.hidden_units = hidden_units
        self.batch_size = batch_size
        self.features_key = features_key
        self.weight_key = weight_key
        self.weight_class_M = weight_class_M

    def fit(self, X, y, num_epochs=1, warm_start=True):
        warm_start = self.check_warm_start(warm_start)
        if not warm_start:
            X, y = self.fit_and_apply_transformers(X, y)
            self.classifier_ = self.create_dnn_classifier()
        else:
            X, y = self.apply_transformers(X, y)

        self.classifier_.train(self.input_fn(
            tf.estimator.ModeKeys.TRAIN, X, y, num_epochs))
        return self

    def fit_and_apply_transformers(self, X, y):
        # Fit and transform X
        self.vectorizer_ = CountVectorizer(lowercase=False,
                                           tokenizer=self.tokens_to_str,
                                           ngram_range=(1, 2))
        X = self.vectorizer_.fit_transform(X)
        self.n_features_ = len(self.vectorizer_.vocabulary_)

        # Fit and transform y
        self.label_encoder_ = LabelEncoder()
        y = self.label_encoder_.fit_transform(y)
        self.n_classes_ = len(self.label_encoder_.classes_)

        return X, y

    def apply_transformers(self, X, y):
        X = self.vectorizer_.transform(X)
        y = self.label_encoder_.transform(y)
        return X, y

    def predict(self, X):
        X = self.vectorizer_.transform(X)
        classes = list(self.classifier_.predict(self.input_fn(
            tf.estimator.ModeKeys.PREDICT, X)))
        labels = self.label_encoder_.inverse_transform(classes)
        return labels
    
    def score(self, X, y):
        X, y = self.apply_transformers(X, y)
        results = self.classifier_.evaluate(self.input_fn(
            tf.estimator.ModeKeys.EVAL, X, y))
        return results["f1-score"]

    def create_dnn_classifier(self):
        # Columns of X
        self.feature_columns_ = [tf.feature_column.numeric_column(
            key=self.features_key, shape=self.n_features_
        )]

        # Column of weights which will be used to compute loss
        weight_column = tf.feature_column.numeric_column(self.weight_key)
        
        # Model parameters
        params = {
            "feature_columns": self.feature_columns_,
            "weight_column": weight_column,
            "hidden_units": self.hidden_units,
            "n_classes": self.n_classes_,
        }
        
        # Create model
        model = tf.estimator.Estimator(model_fn=self.model_fn,
                                       model_dir=self.model_dir,
                                       params=params)
        model = tf.contrib.estimator.add_metrics(model, self.f1_score)
        return model
        
    def model_fn(self, features, labels, mode, params):
        # Network
        net = tf.feature_column.input_layer(features, params['feature_columns'])
        for units in params['hidden_units']:
            net = tf.layers.dense(net, units=units, activation=tf.nn.relu)
        logits = tf.layers.dense(net, params['n_classes'], activation=None)
        
        # Predict
        predicted_classes = tf.argmax(logits, 1)
        if mode == tf.estimator.ModeKeys.PREDICT:
            return tf.estimator.EstimatorSpec(mode, predictions=predicted_classes)        
        
        # Loss
        weights = tf.feature_column.input_layer(features, params['weight_column'])
        loss = tf.losses.sparse_softmax_cross_entropy(labels=labels, 
                                                      logits=logits,
                                                      weights=weights)

        # Eval
        if mode == tf.estimator.ModeKeys.EVAL:
            return tf.estimator.EstimatorSpec(mode, loss=loss, predictions=predicted_classes)
        
        # Train
        optimizer = tf.train.AdagradOptimizer(learning_rate=0.1)
        train_op = optimizer.minimize(loss, global_step=tf.train.get_global_step())
        return tf.estimator.EstimatorSpec(mode, loss=loss, train_op=train_op)
    
    def f1_score(self, labels, predictions):
        return {"f1-score": self.f1_metric_fn(labels=labels, predictions=predictions)}
    
    def f1_metric_fn(self, labels, predictions):
        p, p_op = tf.metrics.precision(labels=labels, predictions=predictions)
        r, r_op = tf.metrics.recall(labels=labels, predictions=predictions)
        return 2 * p * r / (p + r), tf.group(p_op, r_op)        

    def input_fn(self, mode, X, y=None, num_epochs=1):
        n = X.shape[0]
        if mode == tf.estimator.ModeKeys.TRAIN:
            num_batches = n // self.batch_size
        else:
            num_batches = int(np.ceil(n / self.batch_size))

        def mode_input_fn():
            # Convert to sparse tensor
            tf_X = self.convert_sparse_matrix_to_sparse_tensor(X)

            # Batch iterator
            i = tf.train.range_input_producer(
                limit=num_batches, num_epochs=num_epochs, shuffle=False
            ).dequeue()

            # Slice sparse tensor using batch number, then convert to dense
            tf_X = tf.sparse.to_dense(tf.sparse.slice(
                tf_X, start=[i * self.batch_size, 0],
                size=[self.batch_size, self.n_features_]
            ), validate_indices=False)

            if mode == tf.estimator.ModeKeys.PREDICT:
                return {self.features_key: tf_X}
            
            # Slice labels tensor using batch number
            tf_y = tf.convert_to_tensor(y)
            tf_y = tf_y[i * self.batch_size:(i + 1) * self.batch_size]

            # Create weights tensor using given weight of class "M"
            class_M = self.label_encoder_.transform(["M"])
            weights = tf.cast(tf.equal(tf_y, class_M), tf.float64)
            weights = tf.multiply(weights, (self.weight_class_M - 1)) + 1

            # Return tensors in correct format
            return {self.features_key: tf_X, self.weight_key: weights}, tf_y

        return mode_input_fn

    def check_warm_start(self, warm_start):
        if warm_start:
            # Check if model was already fitted
            try:
                self.classifier_
            except:
                warm_start = False
        return warm_start

    def set_model_directory(self, checkpoints_dir, model_name):
        if model_name is not None:
            self.model_dir = checkpoints_dir + model_name
            # Check model_dir doesn't already exist
            if os.path.exists(self.model_dir):
                raise ValueError("model_dir already exists")
        else:
            self.model_dir = None

    @staticmethod
    def tokens_to_str(tokens):
        # Used in CountVectorizer because we already have tokens
        return list(map(str, tokens))

    @staticmethod
    def convert_sparse_matrix_to_sparse_tensor(X):
        coo = X.tocoo()
        indices = np.mat([coo.row, coo.col]).transpose()
        return tf.SparseTensorValue(indices, coo.data, coo.shape)

tf.logging.set_verbosity(tf.logging.ERROR)

num_epochs = 5
weight_class_M = Counter(y_train)["C"] / Counter(y_train)["M"]
model = DNNModel(weight_class_M=weight_class_M, hidden_units=(10,))

for epoch in range(num_epochs):
    model.fit(X_train, y_train, num_epochs=1, warm_start=True)
    f1 = model.score(X_test, y_test)
    print("-" * 80 + "\nEPOCH %d: test f1-score = %.3f" % (epoch, f1))
