==> notebooks/3_embeddings.ipynb <==
import os
import re
import pickle
import numpy as np
import pandas as pd

from collections import Counter
from gensim.models import KeyedVectors

from sklearn.metrics import f1_score, make_scorer, confusion_matrix, \
    classification_report
from sklearn.base import BaseEstimator, ClassifierMixin
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB, BernoulliNB
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, cross_val_score, \
    StratifiedShuffleSplit, RandomizedSearchCV

# Load data

with open("../Data/Learn/sequences.pkl", "rb") as f:
    learn_sequences = pickle.load(f)
    
with open("../Data/Test/sequences.pkl", "rb") as f:
    test_sequences = pickle.load(f)
    
with open("../Data/dict.pkl", "rb") as f:
    vocabulary = pickle.load(f)

# Embeddings

### Load fasttext embeddings

embeddings_dict = KeyedVectors.load_word2vec_format("../Data/wiki.fr.vec")

### Handle unknown words

new_vocab_mapping, new_reverse_vocabulary = {}, {}

# Known words
i = -1
for word in vocabulary:
    if word in embeddings_dict:
        i += 1
        new_vocab_mapping[vocabulary[word]] = i
        new_reverse_vocabulary[i] = word
print("%d known words out of %d" % (i + 1, len(vocabulary)))

# Unknown words
i += 1
for word in vocabulary:
    if word not in embeddings_dict:
        new_vocab_mapping[vocabulary[word]] = i
new_reverse_vocabulary[i] = "<UNK>"

print("%d original words" % len(new_vocab_mapping))
print("%d new words (1 unknown)" % len(np.unique(list(new_vocab_mapping.values()))))
print("%d max new words index" % max(new_vocab_mapping.values()))

num_new_words = 28935
new_reverse_vocabulary[num_new_words - 1]

### Apply new mappings to sequences

new_learn_sequences = [list(map(new_vocab_mapping.get, seq)) for seq in learn_sequences]
new_test_sequences = [list(map(new_vocab_mapping.get, seq)) for seq in test_sequences]

### Create corresponding embeddings array

embeddings = np.zeros((num_new_words, 300))
# Last row is for unknown words, we leave it to zero for the moment
for i in range(num_new_words - 1):
    embeddings[i, :] = embeddings_dict[new_reverse_vocabulary[i]]

### Verification

(learn_sentences[0], new_learn_sequences[0][:5], 
(embeddings[2147] == embeddings_dict.word_vec("aurai")).all())

(test_sentences[0], new_test_sequences[0][:5], 
(embeddings[48] == embeddings_dict.word_vec("tous")).all())

# Last row for unknown words
embeddings[num_new_words - 1]

### Save new data

with open("../Data/generated/my_learn_sequences.pkl", "wb") as f:
    pickle.dump(new_learn_sequences, f)
    
with open("../Data/generated/my_test_sequences.pkl", "wb") as f:
    pickle.dump(new_test_sequences, f)
    
with open("../Data/generated/my_reverse_vocabulary.pkl", "wb") as f:
    pickle.dump(new_reverse_vocabulary, f)
    
with open("../Data/generated/my_embeddings.pkl", "wb") as f:
    pickle.dump(embeddings, f)
